test: consolidate scraper contracts
