--- tests/conftest.py
@@ -1,6 +1,0 @@
-import os
-import sys
-
-# Add the root directory of the project to sys.path
-PROJECT_ROOT = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-sys.path.insert(0, PROJECT_ROOT)
